+++ tests/testthat/test-variable-catalog.R
@@ -1,26 +1,22 @@
 context("Variable catalog")
 
 with_mock_HTTP({
-    variables.catalog.url <- "/api/datasets/dataset1/variables/"
-    varblob <- crGET(variables.catalog.url)
+    ds <- loadDataset("test ds")
+    varcat <- allVariables(ds)
+    varorder <- ordering(varcat)
 
     test_that("VariableCatalog instantiates from Shoji", {
-        expect_is(VariableCatalog(varblob), "VariableCatalog")
+        expect_is(varcat, "VariableCatalog")
     })
-
-    varcat <- VariableCatalog(varblob)
-    order.url <- "/api/datasets/dataset1/variables/hierarchical/"
-    varorder <- VariableOrder(crGET(order.url))
 
     test_that("VariableCatalog index method", {
         expect_identical(names(index(varcat)), names(varcat@index))
-        expect_identical(names(index(varcat)), names(varblob$index))
     })
 
     test_that("VariableCatalog has the right contents", {
         expect_true(all(grepl("/api/datasets/dataset1/variables",
             urls(varcat))))
-        expect_identical(self(varcat), variables.catalog.url)
+        expect_identical(self(varcat), "/api/datasets/dataset1/variables/")
         expect_identical(entities(ordering(varcat)), entities(varorder))
     })
 
@@ -28,7 +24,7 @@ with_mock_HTTP({
         expect_identical(index(active(varcat)),
             index(varcat)[urls(ordering(varcat))])
         expect_equivalent(index(hidden(varcat)), list())
-        index(varcat)[[2]]$discarded <- TRUE
+        index(varcat)[[1]]$discarded <- TRUE
         expect_is(active(varcat), "VariableCatalog")
         expect_is(hidden(varcat), "VariableCatalog")
         expect_identical(urls(active(varcat)),
@@ -75,32 +71,32 @@ with_mock_HTTP({
 
     test_that("attribute getters", {
         expect_identical(names(varcat)[1:3],
-            c("Gender", "Birth Year", "starttime"))
-        expect_identical(aliases(varcat)[1:2], c("gender", "birthyr"))
+            c("Birth Year", "Gender", "mymrset"))
+        expect_identical(aliases(varcat)[1:2], c("birthyr", "gender"))
         expect_identical(types(varcat)[1:3],
-            c("categorical", "numeric", "datetime"))
+            c("numeric", "categorical", "multiple_response"))
         expect_identical(descriptions(varcat[1:3]),
-            c("Gender", NA, "Interview Start Time"))
+            c(NA, "Gender", "Please select all that apply"))
         expect_identical(notes(varcat[1:3]),
-            c("", "Asked instead of age", ""))
+            c("Asked instead of age", "", ""))
     })
 
     test_that("attribute setters", {
-        expect_PATCH(names(varcat)[1:3] <- c("Gender", "Year of birth", "Start time"),
+        expect_PATCH(names(varcat)[1:3] <- c("Year of birth", "Gender", "Start time"),
             "/api/datasets/dataset1/variables/",
             '{"/api/datasets/dataset1/variables/birthyr/":{"name":"Year of birth"},',
-            '"/api/datasets/dataset1/variables/starttime/":{"name":"Start time"}}')
-        expect_PATCH(notes(varcat)[1:3] <- c("", "Asked instead of age", "ms"),
+            '"/api/datasets/dataset1/variables/mymrset/":{"name":"Start time"}}')
+        expect_PATCH(notes(varcat)[1:3] <- c("Asked instead of age", "", "ms"),
             "/api/datasets/dataset1/variables/",
-            '{"/api/datasets/dataset1/variables/starttime/":{"notes":"ms"}}')
+            '{"/api/datasets/dataset1/variables/mymrset/":{"notes":"ms"}}')
     })
 
     test_that("show method", {
         expect_output(varcat[1:3],
             get_output(data.frame(
-                alias=c("gender", "birthyr", "starttime"),
-                name=c("Gender", "Birth Year", "starttime"),
-                type=c("categorical", "numeric", "datetime")
+                alias=c("birthyr", "gender", "mymrset"),
+                name=c("Birth Year", "Gender", "mymrset"),
+                type=c("numeric", "categorical", "multiple_response")
             )))
     })
 })